# Conflicts:
#	BackEnd/park_management/src/main/java/park/management/com/vn/controller/StaffAssignmentController.java
#	BackEnd/park_management/src/main/java/park/management/com/vn/entity/BranchStaff.java
#	BackEnd/park_management/src/main/java/park/management/com/vn/entity/Shift.java
#	BackEnd/park_management/src/main/java/park/management/com/vn/mapper/StaffAssignmentMapper.java
#	BackEnd/park_management/src/main/java/park/management/com/vn/model/request/StaffAssignmentRequest.java
#	BackEnd/park_management/src/main/java/park/management/com/vn/service/ParkBranchService.java
#	BackEnd/park_management/src/main/java/park/management/com/vn/service/TicketService.java

+++ BackEnd/park_management/src/main/java/park/management/com/vn/controller/ParkBranchController.java
@@ -35,13 +35,13 @@ public class ParkBranchController {
 
     @PutMapping("/{id}")
     public ResponseEntity<ParkBranchResponse> update(
-            @PathVariable Integer id,
+            @PathVariable Long id,
             @RequestBody ParkBranchRequest request) {
         return ResponseEntity.ok(service.update(id, request));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         service.delete(id);
         return ResponseEntity.noContent().build();
     }
